unity-speeeeeeeeeed 🚤😉 changed data type to int instead of uint8

+++ src/include/payloads.hpp
@@ -114,7 +114,7 @@ struct heading_payload
 struct speed_payload
 {
 	uint8_t  id;
-	uint8_t  speed;
+	int  speed;
 	time_stamp _last_time_stamp;
 
 	//uint8_t time_value[sizeof time_stamp];
